fixed incorrect accuracy on source validation data
When doing domain mapping, to evaluate we must first map from source to
target domain. This was done for training data since train_step_cyclegan
returns the mapped data. However, for evaluation data, the model is run
in metrics.py, so we need to pass in mapping_model and map if
domain=="source".